criação da pagina de configurações de estoque para ajustar os limites de maximo e minimo em estoque, contento assim um filtro manipulável, mudança do icone de confirmação de produto de venda, adição do cnpj no cadastro de cliente, para que poça ser tanto PF quanto PJ, modificação no banco de dados para aceitar a mudança, mudança de validadores, eudnça no email para aceitar cnpj e cpf alem de outras telas que tinham a opção de cpf para aceitar agr o cnpj também. Isso do que foi solicitado. Modificações a parte são: Ajueste na deleção de produtos já vendido para n haver problema fiscal, colocação de icone de home em todas a paginas para que possa voltar mais rápido para a tela inicial, modificação de theas e topbar

+++ faz_seu_nome/src/database/produtoRepository.ts
@@ -54,6 +54,14 @@ export async function updateProduto(
 
 export async function deleteProduto(id:number) {
   const db = await getDB()
+  const vinculo = await db.getFirstAsync(
+    "SELECT id FROM itens_venda WHERE produto_id = ? LIMIT 1",
+    [id]
+  ) as any;
+
+  if (vinculo) {
+    throw new Error("PRODUTO_COM_VENDAS");
+  }
   await db.runAsync(
     "DELETE FROM produtos WHERE id = ?", [id]
   )
